requirements.txt
dj-database-url==0.5.0
Django==3.0.2
django-cors-headers==3.2.1
django-filter==2.2.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
djangorestframework-recursive==0.1.2
drf-yasg==1.17.1
flex==6.14.1
